vm_dump.c: suppress warning
* vm_dump.c (rb_vmdebug_debug_print_pre): suppress warning on a
  platform ptrdiff_t is shorter than long int.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@43931 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ vm_dump.c
@@ -362,7 +362,7 @@ rb_vmdebug_debug_print_pre(rb_thread_t *th, rb_control_frame_t *cfp,VALUE *_pc)
 	    printf(" ");
 	}
 	printf("| ");
-	if(0)printf("[%03ld] ",cfp->sp - th->stack);
+	if(0)printf("[%03ld] ", (long)(cfp->sp - th->stack));
 
 	/* printf("%3"PRIdPTRDIFF" ", VM_CFP_CNT(th, cfp)); */
 	if (pc >= 0) {
